+++ src/components/ChangePasswordModal.js
@@ -1,5 +1,5 @@
 import { useState,useEffect } from "react";
-import Modal from './Modal'
+import Modal from './modal/Modal'
 
 const ChangePasswordModal = ({
     isOpen,
