+++ src/components/CreateFolderModal.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Modal from './Modal';
+import Modal from './modal/Modal';
 
 function CreateFolderModal({ 
     isOpen,
