+++ src/components/modal/user/DeleteUserModal.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Modal from '../../Modal';
+import Modal from '../Modal';
 
 function DeleteUserPasswordModal({
 	isOpen,
