+++ src/components/modal/user/UpdateUserPasswordModal.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Modal from '../../Modal';
+import Modal from '../Modal';
 
 import UpdatePasswordRequest from '../../../entity/UpdatePasswordRequest';
